requirements.txt
asgiref==3.2.10
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.3
channels==3.0.1
chardet==3.0.4
constantly==15.1.0
cryptography==3.2.1
daphne==3.0.0
Django==3.0.10
django-phonenumber-field==5.0.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
phonenumbers==8.12.11
Pillow==8.0.1
pkg-resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pytz==2020.1
requests==2.24.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
twilio==6.46.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.11
zope.interface==5.2.0
